Some integrated tests work, but still need to be able to create files
Having trouble with forgoing traditional DynamoDB config in the app.config. Looking at
source code for AWS on github looks like there's still a lot of stuff going on with configuration
as the AWS team tries to figure out .net standard and .net core. I'll come back to this in a bit
when the configuration story for .net standard improves.